feat: channel 로직 재사용을 위해 컴포저블 방식 도입
